requirements.txt
Flask==2.0.1
Flask-JWT-Extended==4.2.3
Flask-Cors==3.0.10
Flask-Bcrypt==0.7.1
psycopg2==2.9.1
SQLAlchemy==1.4.23
python-dotenv==0.19.0
gunicorn==21.2.0
stripe==2.60.0
Werkzeug==2.0.1
